refactor store types into models
